feat(#62): 홈, 로그인 경로 변경
