+++ src/app/page.tsx
@@ -1,240 +1,117 @@
 "use client";
 
-import { postLogin } from "@/api/login";
-import Button from "@/components/common/Button";
-import Checkbox from "@/components/common/Checkbox";
-import Input from "@/components/common/Input";
-import { emailRegEx } from "@/constants/regEx";
-import { setUserName, setUserProfileImg } from "@/redux/slices/userSlice";
-import { theme } from "@/styles/theme";
-import { setToken } from "@/utils/storage";
-import { AxiosError } from "axios";
-import Image from "next/image";
-import Link from "next/link";
+import ChatButton from "@/components/common/ChatButton";
 import { useRouter } from "next/navigation";
-import { useState } from "react";
-import { useDispatch } from "react-redux";
+import { MATCH_PAGE_DATA } from "@/data/match";
+import Image from "next/image";
 import styled from "styled-components";
+import { theme } from "@/styles/theme";
 
-const Login = () => {
+const HomePage = () => {
   const router = useRouter();
-  const dispatch = useDispatch();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  const [emailValid, setEmailValid] = useState<boolean | undefined>(undefined);
-  const [passwordValid, setPasswordValid] = useState<boolean | undefined>(
-    undefined
-  );
-  const [autoLogin, setAutoLogin] = useState(false);
-
-  const validateEmail = (email: string) => {
-    setEmailValid(emailRegEx.test(email));
-  };
-
-  const validatePassword = (password: string) => {
-    if (password.length === 0) {
-      setPasswordValid(undefined);
-    } else {
-      setPasswordValid(true);
-    }
-  };
-
-  /* 로그인 */
-  const handleLogin = async () => {
-    try {
-      const response = await postLogin({ email, password });
-      const accessToken = response.result.accessToken;
-      const refreshToken = response.result.refreshToken;
-
-      /* 자동 로그인 체크 여부에 따라 토큰 저장 위치 결정 */
-      setToken(accessToken, refreshToken, autoLogin);
-
-      dispatch(setUserName(response.result.name));
-      dispatch(setUserProfileImg(response.result.profileImage));
-      router.push("/home");
-    } catch (error) {
-      if (error instanceof AxiosError) {
-        const status = error.response?.status;
-        if (status === 401) {
-          // 401 Unauthorized 처리
-          setPasswordValid(false);
-        } else {
-          // 기타 에러 처리
-          setEmailValid(false);
-          setPasswordValid(false);
-        }
-      } else {
-        // 예상치 못한 에러 처리
-        setEmailValid(false);
-        setPasswordValid(false);
-      }
-    }
-  };
 
   return (
-    <Container>
-      <Box>
-        <Image
-          src="/assets/icons/logo_m.svg"
-          width={277}
-          height={88}
-          alt="logo"
-        />
-        <Title>로그인</Title>
-        <P>GAMGOO에 오신 것을 환영합니다.</P>
-        <Content>
-          <Div>
-            <InputList>
-              <Input
-                inputType="input"
-                value={email}
-                onChange={(value) => {
-                  setEmail(value);
-                  validateEmail(value);
-                }}
-                placeholder="이메일 주소"
-                isValid={emailValid}
-              />
-              <Input
-                inputType="password"
-                value={password}
-                onChange={(value) => {
-                  setPassword(value);
-                  validatePassword(value);
-                }}
-                placeholder="비밀번호"
-                isValid={passwordValid}
-              />
-            </InputList>
-            <Button
-              buttonType="primary"
-              text="이메일로 시작하기"
-              onClick={handleLogin}
-              disabled={!email || !password || !emailValid || !passwordValid}
-            />
-          </Div>
-          <Check>
-            <P>
-              <Checkbox
-                value="autoLogin"
-                isChecked={autoLogin}
-                onChange={(isChecked) => setAutoLogin(isChecked)}
-                gap="0px"
-              />
-              자동 로그인
-              <Bar />
-              <Link href="/password/find">비밀번호 찾기</Link>
-            </P>
-          </Check>
-        </Content>
-        <Line />
-        <SocialIcons>
+    <Wrapper>
+      <HomeContent>
+        <Header>
           <Image
-            src="/assets/icons/naver.svg"
-            width={58}
-            height={58}
-            alt="naver"
+            src="/assets/icons/logo_m.svg"
+            width={371}
+            height={117}
+            priority
+            alt="logo"
           />
-          <Image
-            src="/assets/icons/kakao.svg"
-            width={58}
-            height={58}
-            alt="kakao"
-          />
-          <Image
-            src="/assets/icons/google.svg"
-            width={58}
-            height={58}
-            alt="google"
-          />
-        </SocialIcons>
-        <P>
-          아직 GAMMGOO 회원이 아니신가요?{`   `}
-          <Join href="/join/terms">회원가입</Join>
-        </P>
-      </Box>
-    </Container>
+          <SubTitle>겜구 커뮤니티에 오신 것을 환영합니다.</SubTitle>
+        </Header>
+        <Main>
+          {MATCH_PAGE_DATA.map((content) => {
+            return (
+              <ContentWrapper
+                key={content.id}
+                onClick={() => router.push(content.pathname)}
+              >
+                <Image
+                  src={content.image}
+                  width={0}
+                  height={0}
+                  style={{ width: "100%", height: "auto" }}
+                  alt={content.title}
+                  priority
+                />
+                <ContentTitle>{content.title}</ContentTitle>
+              </ContentWrapper>
+            );
+          })}
+        </Main>
+        <Footer>
+          <ChatBoxContent>
+            <ChatButton count={3} />
+          </ChatBoxContent>
+        </Footer>
+      </HomeContent>
+    </Wrapper>
   );
 };
 
-export default Login;
+export default HomePage;
 
-const Container = styled.div`
+const Wrapper = styled.div`
   width: 100%;
-  height: 100vh;
   display: flex;
   justify-content: center;
-  align-items: center;
+  padding-top: 140px;
 `;
 
-const Box = styled.div`
-  max-width: 540px;
+const HomeContent = styled.div`
+  max-width: 1440px;
   width: 100%;
-  padding: 35px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+  padding: 0px 80px;
 `;
 
-const Title = styled.div`
-  margin-top: 26px;
+const Header = styled.header`
+  margin-bottom: 35px;
+`;
+
+const SubTitle = styled.div`
+  ${(props) => props.theme.fonts.regular25};
   color: #44515c;
-  ${(props) => props.theme.fonts.regular35};
 `;
 
-const Content = styled.div`
-  width: 100%;
-  margin-top: 49px;
-`;
-
-const P = styled.div`
-  display: flex;
-  color: ${theme.colors.gray200};
-  ${(props) => props.theme.fonts.regular14};
-  gap: 10px;
-`;
-
-const Div = styled.div`
+const Main = styled.main`
   width: 100%;
   display: flex;
-  flex-direction: column;
-  gap: 17px;
+  justify-content: center;
+  align-items: center;
+  gap: 59px;
+  margin-bottom: 37px;
+
+  @media (max-width: 1200px) {
+    flex-direction: column;
+    justify-content: center;
+  }
+`;
+
+const ContentWrapper = styled.div`
+  max-width: 600px;
   position: relative;
+  cursor: pointer;
 `;
 
-const InputList = styled(Div)`
-  gap: 8px;
+const ContentTitle = styled.p`
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  ${(props) => props.theme.fonts.bold32};
+  color: ${theme.colors.white};
+  white-space: nowrap;
 `;
 
-const Check = styled.div`
+const Footer = styled.footer`
   display: flex;
-  justify-content: flex-end;
-  margin-top: 22px;
+  margin-bottom: 78px;
 `;
 
-const Bar = styled.div`
-  width: 1px;
-  height: 20px;
-  background: ${theme.colors.gray300};
-`;
-
-const Line = styled.div`
-  width: 100%;
-  height: 1px;
-  background: #d4d4d4;
-  margin: 54px 0 74px 0;
-`;
-
-const SocialIcons = styled.div`
-  width: 250px;
-  display: flex;
-  justify-content: space-around;
-  margin-bottom: 57px;
-`;
-
-const Join = styled(Link)`
-  color: ${theme.colors.gray300};
-  font-weight: 500;
+const ChatBoxContent = styled.div`
+  margin-left: auto;
 `;
