+++ src/components/createBoard/PostBoard.tsx
@@ -229,7 +229,7 @@ const PostBoard = (props: PostBoardProps) => {
           height={58}
           content="로그아웃 되었습니다. 다시 로그인 해주세요."
           alt="로그인 필요"
-          onClose={() => router.push("/")}
+          onClose={() => router.push("/login")}
         />
       )}
